Changed debouncing method of the detector to work more reliably.
Also added the possibility to reset the state machine by holding down both direction buttons.

Added a 1000ms delay before measuring current when the train is departing.

TODO:
if a train takes too long to depart, it is still drawing current but is stuck. The power must be cut off and the state machine must be reset.
if a train takes too long to arrive at the station Leds need to blink to indicate this.
There also needs to be a blinking action for when there is a short circuit.
The automated restart of the short circuit code should be replaced by the manual restart